nvmesh to es6 class, split out loaders and utils
